extract type to own file
