+++ homeassistant/components/vicare/types.py
@@ -1,0 +1,22 @@
+"""Types for the ViCare integration."""
+from __future__ import annotations
+
+from collections.abc import Callable
+from dataclasses import dataclass
+from typing import Any
+
+from PyViCare.PyViCareDevice import Device
+
+
+@dataclass(frozen=True)
+class ViCareRequiredKeysMixin:
+    """Mixin for required keys."""
+
+    value_getter: Callable[[Device], Any]
+
+
+@dataclass(frozen=True)
+class ViCareRequiredKeysMixinWithSet(ViCareRequiredKeysMixin):
+    """Mixin for required keys with setter."""
+
+    value_setter: Callable[[Device], bool]
